requirements.txt
decorator==4.1.2
Django==1.11
django-audit-log==0.7.0
djangorestframework==3.6.4
factory-boy==2.9.2
Faker==0.8.8
ipython==6.2.1
ipython-genutils==0.2.0
jedi==0.11.1
parso==0.1.1
pexpect==4.3.1
pickleshare==0.7.4
Pillow==5.0.0
prompt-toolkit==1.0.15
ptyprocess==0.5.2
Pygments==2.2.0
python-dateutil==2.6.1
pytz==2017.3
simplegeneric==0.8.1
six==1.11.0
text-unidecode==1.1
traitlets==4.3.2
typing==3.6.2
wcwidth==0.1.7
wheel==0.24.0
